+++ cmd/harp/internal/cmd/bug.go
@@ -1,19 +1,7 @@
-// Licensed to Elasticsearch B.V. under one or more contributor
-// license agreements. See the NOTICE file distributed with
-// this work for additional information regarding copyright
-// ownership. Elasticsearch B.V. licenses this file to you under
-// the Apache License, Version 2.0 (the "License"); you may
-// not use this file except in compliance with the License.
-// You may obtain a copy of the License at
+// SPDX-FileCopyrightText: 2019 Elasticsearch B.V.
+// SPDX-FileCopyrightText: 2019-2023 Thibault NORMAND <me@zenithar.org>
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
+// SPDX-License-Identifier: Apache-2.0 AND MIT
 
 package cmd
 
@@ -25,8 +13,8 @@ import (
 	"github.com/skratchdot/open-golang/open"
 	"github.com/spf13/cobra"
 
-	"github.com/zntrio/harp/v2/pkg/sdk/cmdutil"
-	"github.com/zntrio/harp/v2/pkg/sdk/log"
+	"zntr.io/harp/v2/pkg/sdk/cmdutil"
+	"zntr.io/harp/v2/pkg/sdk/log"
 )
 
 // -----------------------------------------------------------------------------
@@ -56,9 +44,9 @@ func runBug(cmd *cobra.Command, args []string) {
 	body := cmdutil.BugReport()
 
 	// Open the browser to issue creation form
-	reportURL := "https://github.com/zntrio/harp/issues/new?body=" + url.QueryEscape(body)
+	reportURL := "https://zntr.io/harp/issues/new?body=" + url.QueryEscape(body)
 	if err := open.Run(reportURL); err != nil {
-		fmt.Fprint(os.Stdout, "Please file a new issue at github.com/zntrio/harp/issues/new using this template:\n\n")
+		fmt.Fprint(os.Stdout, "Please file a new issue at zntr.io/harp/issues/new using this template:\n\n")
 		fmt.Fprint(os.Stdout, body)
 	}
 }