+++ cmd/harp/loader_nonfips.go
@@ -1,19 +1,7 @@
-// Licensed to Elasticsearch B.V. under one or more contributor
-// license agreements. See the NOTICE file distributed with
-// this work for additional information regarding copyright
-// ownership. Elasticsearch B.V. licenses this file to you under
-// the Apache License, Version 2.0 (the "License"); you may
-// not use this file except in compliance with the License.
-// You may obtain a copy of the License at
+// SPDX-FileCopyrightText: 2019 Elasticsearch B.V.
+// SPDX-FileCopyrightText: 2019-2023 Thibault NORMAND <me@zenithar.org>
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
+// SPDX-License-Identifier: Apache-2.0 AND MIT
 
 //go:build !fips
 
@@ -28,18 +16,18 @@ import (
 	_ "crypto/sha512"
 
 	// Register encryption transformers.
-	_ "github.com/zntrio/harp/v2/pkg/sdk/value/encryption/aead"
-	_ "github.com/zntrio/harp/v2/pkg/sdk/value/encryption/age"
-	_ "github.com/zntrio/harp/v2/pkg/sdk/value/encryption/branca"
-	_ "github.com/zntrio/harp/v2/pkg/sdk/value/encryption/dae"
-	_ "github.com/zntrio/harp/v2/pkg/sdk/value/encryption/fernet"
-	_ "github.com/zntrio/harp/v2/pkg/sdk/value/encryption/jwe"
-	_ "github.com/zntrio/harp/v2/pkg/sdk/value/encryption/paseto"
-	_ "github.com/zntrio/harp/v2/pkg/sdk/value/encryption/secretbox"
-	_ "github.com/zntrio/harp/v2/pkg/sdk/value/signature/jws"
-	_ "github.com/zntrio/harp/v2/pkg/sdk/value/signature/paseto"
-	_ "github.com/zntrio/harp/v2/pkg/sdk/value/signature/raw"
-	_ "github.com/zntrio/harp/v2/pkg/vault"
+	_ "zntr.io/harp/v2/pkg/sdk/value/encryption/aead"
+	_ "zntr.io/harp/v2/pkg/sdk/value/encryption/age"
+	_ "zntr.io/harp/v2/pkg/sdk/value/encryption/branca"
+	_ "zntr.io/harp/v2/pkg/sdk/value/encryption/dae"
+	_ "zntr.io/harp/v2/pkg/sdk/value/encryption/fernet"
+	_ "zntr.io/harp/v2/pkg/sdk/value/encryption/jwe"
+	_ "zntr.io/harp/v2/pkg/sdk/value/encryption/paseto"
+	_ "zntr.io/harp/v2/pkg/sdk/value/encryption/secretbox"
+	_ "zntr.io/harp/v2/pkg/sdk/value/signature/jws"
+	_ "zntr.io/harp/v2/pkg/sdk/value/signature/paseto"
+	_ "zntr.io/harp/v2/pkg/sdk/value/signature/raw"
+	_ "zntr.io/harp/v2/pkg/vault"
 
 	_ "golang.org/x/crypto/blake2b"
 	_ "golang.org/x/crypto/blake2s"