+++ pkg/bundle/ruleset/ruleset_test.go
@@ -1,19 +1,7 @@
-// Licensed to Elasticsearch B.V. under one or more contributor
-// license agreements. See the NOTICE file distributed with
-// this work for additional information regarding copyright
-// ownership. Elasticsearch B.V. licenses this file to you under
-// the Apache License, Version 2.0 (the "License"); you may
-// not use this file except in compliance with the License.
-// You may obtain a copy of the License at
+// SPDX-FileCopyrightText: 2019 Elasticsearch B.V.
+// SPDX-FileCopyrightText: 2019-2023 Thibault NORMAND <me@zenithar.org>
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
+// SPDX-License-Identifier: Apache-2.0 AND MIT
 
 package ruleset
 
@@ -21,7 +9,7 @@ import (
 	"reflect"
 	"testing"
 
-	bundlev1 "github.com/zntrio/harp/v2/api/gen/go/harp/bundle/v1"
+	bundlev1 "zntr.io/harp/v2/api/gen/go/harp/bundle/v1"
 )
 
 func TestValidate(t *testing.T) {
@@ -50,7 +38,7 @@ func TestValidate(t *testing.T) {
 			name: "invalid kind",
 			args: args{
 				spec: &bundlev1.RuleSet{
-					ApiVersion: "harp.elastic.co/v1",
+					ApiVersion: "harp.zntr.io/v2",
 					Kind:       "foo",
 				},
 			},
@@ -60,7 +48,7 @@ func TestValidate(t *testing.T) {
 			name: "nil meta",
 			args: args{
 				spec: &bundlev1.RuleSet{
-					ApiVersion: "harp.elastic.co/v1",
+					ApiVersion: "harp.zntr.io/v2",
 					Kind:       "RuleSet",
 				},
 			},
@@ -70,7 +58,7 @@ func TestValidate(t *testing.T) {
 			name: "meta name not defined",
 			args: args{
 				spec: &bundlev1.RuleSet{
-					ApiVersion: "harp.elastic.co/v1",
+					ApiVersion: "harp.zntr.io/v2",
 					Kind:       "RuleSet",
 					Meta:       &bundlev1.RuleSetMeta{},
 				},
@@ -81,7 +69,7 @@ func TestValidate(t *testing.T) {
 			name: "nil spec",
 			args: args{
 				spec: &bundlev1.RuleSet{
-					ApiVersion: "harp.elastic.co/v1",
+					ApiVersion: "harp.zntr.io/v2",
 					Kind:       "RuleSet",
 					Meta:       &bundlev1.RuleSetMeta{},
 				},
@@ -92,7 +80,7 @@ func TestValidate(t *testing.T) {
 			name: "no action patch",
 			args: args{
 				spec: &bundlev1.RuleSet{
-					ApiVersion: "harp.elastic.co/v1",
+					ApiVersion: "harp.zntr.io/v2",
 					Kind:       "RuleSet",
 					Meta:       &bundlev1.RuleSetMeta{},
 					Spec:       &bundlev1.RuleSetSpec{},
@@ -128,14 +116,14 @@ func TestChecksum(t *testing.T) {
 			name: "valid",
 			args: args{
 				spec: &bundlev1.RuleSet{
-					ApiVersion: "harp.elastic.co/v1",
+					ApiVersion: "harp.zntr.io/v2",
 					Kind:       "RuleSet",
 					Meta:       &bundlev1.RuleSetMeta{},
 					Spec:       &bundlev1.RuleSetSpec{},
 				},
 			},
 			wantErr: false,
-			want:    "yM_TR6rMWW7BGA1Ms-U3WK6E4Xax5qRBMjK4VQLyZmQ",
+			want:    "WDRJqdZMEM5A5xwUbv4J3K_pNCFHjczR97hVupD7nP4",
 		},
 	}
 	for _, tt := range tests {