+++ pkg/bundle/template/executor_test.go
@@ -1,19 +1,7 @@
-// Licensed to Elasticsearch B.V. under one or more contributor
-// license agreements. See the NOTICE file distributed with
-// this work for additional information regarding copyright
-// ownership. Elasticsearch B.V. licenses this file to you under
-// the Apache License, Version 2.0 (the "License"); you may
-// not use this file except in compliance with the License.
-// You may obtain a copy of the License at
+// SPDX-FileCopyrightText: 2019 Elasticsearch B.V.
+// SPDX-FileCopyrightText: 2019-2023 Thibault NORMAND <me@zenithar.org>
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
+// SPDX-License-Identifier: Apache-2.0 AND MIT
 
 package template
 
@@ -23,9 +11,9 @@ import (
 
 	fuzz "github.com/google/gofuzz"
 
-	bundlev1 "github.com/zntrio/harp/v2/api/gen/go/harp/bundle/v1"
-	"github.com/zntrio/harp/v2/pkg/bundle/template/visitor/secretbuilder"
-	"github.com/zntrio/harp/v2/pkg/template/engine"
+	bundlev1 "zntr.io/harp/v2/api/gen/go/harp/bundle/v1"
+	"zntr.io/harp/v2/pkg/bundle/template/visitor/secretbuilder"
+	"zntr.io/harp/v2/pkg/template/engine"
 )
 
 func TestValidate(t *testing.T) {
@@ -54,7 +42,7 @@ func TestValidate(t *testing.T) {
 			name: "invalid kind",
 			args: args{
 				spec: &bundlev1.Template{
-					ApiVersion: "harp.elastic.co/v1",
+					ApiVersion: "harp.zntr.io/v2",
 					Kind:       "foo",
 				},
 			},
@@ -64,7 +52,7 @@ func TestValidate(t *testing.T) {
 			name: "nil meta",
 			args: args{
 				spec: &bundlev1.Template{
-					ApiVersion: "harp.elastic.co/v1",
+					ApiVersion: "harp.zntr.io/v2",
 					Kind:       "BundleTemplate",
 				},
 			},
@@ -74,7 +62,7 @@ func TestValidate(t *testing.T) {
 			name: "meta name not defined",
 			args: args{
 				spec: &bundlev1.Template{
-					ApiVersion: "harp.elastic.co/v1",
+					ApiVersion: "harp.zntr.io/v2",
 					Kind:       "BundleTemplate",
 					Meta:       &bundlev1.TemplateMeta{},
 				},
@@ -85,7 +73,7 @@ func TestValidate(t *testing.T) {
 			name: "nil spec",
 			args: args{
 				spec: &bundlev1.Template{
-					ApiVersion: "harp.elastic.co/v1",
+					ApiVersion: "harp.zntr.io/v2",
 					Kind:       "BundleTemplate",
 					Meta:       &bundlev1.TemplateMeta{},
 				},
@@ -96,7 +84,7 @@ func TestValidate(t *testing.T) {
 			name: "no action template",
 			args: args{
 				spec: &bundlev1.Template{
-					ApiVersion: "harp.elastic.co/v1",
+					ApiVersion: "harp.zntr.io/v2",
 					Kind:       "BundleTemplate",
 					Meta:       &bundlev1.TemplateMeta{},
 					Spec:       &bundlev1.TemplateSpec{},
@@ -132,14 +120,14 @@ func TestChecksum(t *testing.T) {
 			name: "valid",
 			args: args{
 				spec: &bundlev1.Template{
-					ApiVersion: "harp.elastic.co/v1",
+					ApiVersion: "harp.zntr.io/v2",
 					Kind:       "BundleTemplate",
 					Meta:       &bundlev1.TemplateMeta{},
 					Spec:       &bundlev1.TemplateSpec{},
 				},
 			},
 			wantErr: false,
-			want:    "qnYJsLsuawKi7c3A4mgrOm9akKKdt57NJdRR92xlOfA",
+			want:    "3ipnWuWHabucGE3J1xGS0W5GxWtwTOeiI8kb5PucRkY",
 		},
 	}
 	for _, tt := range tests {
@@ -163,7 +151,7 @@ func TestExecute_Fuzz(t *testing.T) {
 
 		// Prepare arguments
 		spec := &bundlev1.Template{
-			ApiVersion: "harp.elastic.co/v1",
+			ApiVersion: "harp.zntr.io/v2",
 			Kind:       "BundleTemplate",
 			Meta:       &bundlev1.TemplateMeta{},
 			Spec:       &bundlev1.TemplateSpec{},