+++ pkg/tasks/lint/validate.go
@@ -1,19 +1,7 @@
-// Licensed to Elasticsearch B.V. under one or more contributor
-// license agreements. See the NOTICE file distributed with
-// this work for additional information regarding copyright
-// ownership. Elasticsearch B.V. licenses this file to you under
-// the Apache License, Version 2.0 (the "License"); you may
-// not use this file except in compliance with the License.
-// You may obtain a copy of the License at
+// SPDX-FileCopyrightText: 2019 Elasticsearch B.V.
+// SPDX-FileCopyrightText: 2019-2023 Thibault NORMAND <me@zenithar.org>
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
+// SPDX-License-Identifier: Apache-2.0 AND MIT
 
 package lint
 
@@ -26,11 +14,11 @@ import (
 	"github.com/xeipuuv/gojsonschema"
 	"sigs.k8s.io/yaml"
 
-	"github.com/zntrio/harp/v2/pkg/bundle"
-	"github.com/zntrio/harp/v2/pkg/bundle/patch"
-	"github.com/zntrio/harp/v2/pkg/bundle/ruleset"
-	"github.com/zntrio/harp/v2/pkg/bundle/template"
-	"github.com/zntrio/harp/v2/pkg/tasks"
+	"zntr.io/harp/v2/pkg/bundle"
+	"zntr.io/harp/v2/pkg/bundle/patch"
+	"zntr.io/harp/v2/pkg/bundle/ruleset"
+	"zntr.io/harp/v2/pkg/bundle/template"
+	"zntr.io/harp/v2/pkg/tasks"
 )
 
 // ValidateTask implements input linter task.
@@ -84,7 +72,7 @@ func (t *ValidateTask) Run(ctx context.Context) error {
 		}
 
 		// Check API Version
-		if specBody.APIVersion != "harp.elastic.co/v1" {
+		if specBody.APIVersion != "harp.zntr.io/v2" {
 			return fmt.Errorf("unsupported YAML file format %q", specBody.APIVersion)
 		}