+++ pkg/vault/transformer.go
@@ -1,19 +1,7 @@
-// Licensed to Elasticsearch B.V. under one or more contributor
-// license agreements. See the NOTICE file distributed with
-// this work for additional information regarding copyright
-// ownership. Elasticsearch B.V. licenses this file to you under
-// the Apache License, Version 2.0 (the "License"); you may
-// not use this file except in compliance with the License.
-// You may obtain a copy of the License at
+// SPDX-FileCopyrightText: 2019 Elasticsearch B.V.
+// SPDX-FileCopyrightText: 2019-2023 Thibault NORMAND <me@zenithar.org>
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
+// SPDX-License-Identifier: Apache-2.0 AND MIT
 
 package vault
 
@@ -22,12 +10,12 @@ import (
 	"path"
 	"strings"
 
-	"github.com/zntrio/harp/v2/pkg/sdk/value"
-	"github.com/zntrio/harp/v2/pkg/sdk/value/encryption"
-	"github.com/zntrio/harp/v2/pkg/sdk/value/encryption/aead"
-	"github.com/zntrio/harp/v2/pkg/sdk/value/encryption/envelope"
-	"github.com/zntrio/harp/v2/pkg/sdk/value/encryption/secretbox"
-	vaultpath "github.com/zntrio/harp/v2/pkg/vault/path"
+	"zntr.io/harp/v2/pkg/sdk/value"
+	"zntr.io/harp/v2/pkg/sdk/value/encryption"
+	"zntr.io/harp/v2/pkg/sdk/value/encryption/aead"
+	"zntr.io/harp/v2/pkg/sdk/value/encryption/envelope"
+	"zntr.io/harp/v2/pkg/sdk/value/encryption/secretbox"
+	vaultpath "zntr.io/harp/v2/pkg/vault/path"
 )
 
 type DataEncryption string