+++ test/magefile.go
@@ -1,19 +1,7 @@
-// Licensed to Elasticsearch B.V. under one or more contributor
-// license agreements. See the NOTICE file distributed with
-// this work for additional information regarding copyright
-// ownership. Elasticsearch B.V. licenses this file to you under
-// the Apache License, Version 2.0 (the "License"); you may
-// not use this file except in compliance with the License.
-// You may obtain a copy of the License at
+// SPDX-FileCopyrightText: 2019 Elasticsearch B.V.
+// SPDX-FileCopyrightText: 2019-2023 Thibault NORMAND <me@zenithar.org>
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
+// SPDX-License-Identifier: Apache-2.0 AND MIT
 
 //go:build mage
 // +build mage
@@ -23,7 +11,7 @@ package main
 import (
 	"github.com/magefile/mage/mg"
 
-	"github.com/zntrio/harp/v2/build/mage/golang"
+	"zntr.io/harp/v2/build/mage/golang"
 )
 
 // -----------------------------------------------------------------------------
@@ -32,14 +20,14 @@ type Fuzz mg.Namespace
 
 func (Fuzz) BundleLoader() {
 	mg.SerialDeps(
-		golang.FuzzBuild("bundle-loader", "github.com/zntrio/harp/v2/test/fuzz/bundle/loader"),
+		golang.FuzzBuild("bundle-loader", "zntr.io/harp/v2/test/fuzz/bundle/loader"),
 		golang.FuzzRun("bundle-loader"),
 	)
 }
 
 func (Fuzz) TemplateReader() {
 	mg.SerialDeps(
-		golang.FuzzBuild("template-reader", "github.com/zntrio/harp/v2/test/fuzz/template"),
+		golang.FuzzBuild("template-reader", "zntr.io/harp/v2/test/fuzz/template"),
 		golang.FuzzRun("template-reader"),
 	)
 }